requirements.txt
flask>=2.2.0,<3.0.0
flask-cors>=3.0.10,<5.0.0
gunicorn>=20.1.0,<22.0.0
firebase-admin>=6.0.1,<7.0.0
pandas>=2.1.0,<2.3.0
numpy>=1.24.0,<2.0.0
openpyxl>=3.1.0,<3.2.0
python-dotenv>=0.21.0,<1.0.0
scikit-learn>=1.2.0,<1.4.0
joblib>=1.2.0,<2.0.0
statsmodels>=0.13.5,<0.14.0
matplotlib>=3.6.0,<3.8.0
seaborn>=0.12.0,<0.13.0
psycopg2-binary>=2.9.0,<3.0.0

